% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activity}
\alias{activity}
\title{Gene Expression Activity Estimation Using Model Matrix}
\usage{
activity(expr, model, scale = TRUE)
}
\arguments{
\item{expr}{A matrix or data frame containing gene expression data, where rows represent genes and columns represent samples.}

\item{model}{A matrix or data frame containing the model for gene activity (e.g., gene signature or pathway information).}

\item{scale}{A logical value (default = TRUE). If TRUE, scales the resulting activity scores for each sample.}
}
\value{
A matrix of gene activity scores for each sample based on the provided model matrix.
}
\description{
This function calculates gene activity based on a model matrix (such as pathway or gene signature models).
It multiplies the gene expression matrix with the model matrix for a set of common genes.
}
\examples{
ac <- activity(expr, model, scale = FALSE)
}
