% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{cpm}
\alias{cpm}
\title{Calculate Counts Per Million (CPM)}
\usage{
cpm(expr)
}
\arguments{
\item{expr}{A matrix or data frame containing raw count data, where rows represent genes and columns represent samples.}
}
\value{
A matrix with CPM values, where each column is normalized to a sum of 1,000,000.
}
\description{
This function calculates the counts per million (CPM) for a given count matrix.
It normalizes the count data by scaling each column (sample) based on the total count sum for that column.
}
\examples{
expr <- cpm(expr)
}
