% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{cpm_normalized}
\alias{cpm_normalized}
\title{Normalize Counts Matrix to CPM and Apply Log Normalization (Optional)}
\usage{
cpm_normalized(counts.matrix, filter = TRUE, log_normalize = TRUE)
}
\arguments{
\item{counts.matrix}{A matrix or data frame containing raw count data, where rows represent genes and columns represent samples.}

\item{filter}{A logical value (default = TRUE). If TRUE, filters out genes with an average CPM less than 0.25 across samples.}

\item{log_normalize}{A logical value (default = TRUE). If TRUE, applies log transformation to the normalized CPM values (with a small offset to avoid log(0)).}
}
\value{
A matrix of CPM values after normalization and (optionally) log transformation and filtering.
}
\description{
This function normalizes a count matrix to CPM, applies log normalization if specified,
and filters out rows with low mean CPM values across samples if the `filter` option is TRUE.
}
\examples{
expr <- cpm_normalized(expr)
}
